Fix IRC bridge Dockerfile
Something appeared to be wrong with the final COPY instruction.  Either I had
something in my local repo that I was unaware of and never committed, or the
behaviour of Docker has since changed.

This also switches to the common (but by no means universal) Docker convention
of putting the program in /opt/app/